requirements.txt
aiohttp==3.12.15
backports.tarfile==1.2.0
beautifulsoup4==4.14.2
easydict==1.13
einops==0.8.1
ffmpeg-python==0.2.0
gensim==4.3.3
h5py==3.14.0
importlib-metadata==8.0.0
jaraco.collections==5.1.0
loguru==0.7.3
matplotlib==3.8.4
nltk==3.9.2
nnaudio==0.3.3
opencv-python==4.11.0.86
pandas==2.2.3
pip-chill==1.0.3
psutil==7.1.0
pydub==0.25.1
pytorch-lightning==2.5.5
pywavelets==1.9.0
resampy==0.4.3
ruamel.yaml==0.18.15
scikit-image==0.25.2
sentence-transformers==5.1.1
tensorboard==2.20.0
tensorboardx==2.6.4
timm==0.4.5
tomli==2.0.1
torchaudio==2.6.0
torchlibrosa==0.1.0
torchtext==0.6.0
wandb==0.22.1
